+++ CHANGES.md
@@ -49,6 +49,7 @@ Changes in 2.9
 - Add jQuery MiniColors plugin & ColorInput() function in assets/js/jquery-minicolors/ & Inputs.php
 - Add MakeTipMessage() & MakeStudentPhotoTipMessage() functions in ProgramFunctions/TipMessage.fnc.php
 - Add Photo on mouse over Student Name in InputFinalGrades.php & Grades.php
+- Highlight color: add CSS to HTML head in Warehouse.php
 
 Changes in 2.8.11
 -----------------